LayoutTests/imported/w3c/web-platform-tests/css/css-position/position-absolute-table-001-expected.txt

Table css width/height are interpreted differently: they are the minimum width. Absolute positioning code should respect this.

t1
div makes cell larger.
t1
div makes cell larger.
t1
div makes cell larger.

FAIL table size is interpreted as border-box width by default assert_equals: right position is 0 expected 260 but got 300
FAIL table size is interpreted as minimum width assert_equals: right position is 0 expected 360 but got 300
PASS table size border-box
FAIL table size border-box interpreted as minimum width assert_equals: right position is 0 expected 400 but got 300
FAIL table size content-box assert_equals: expected 100 but got 60
FAIL table size content-box interpreted as minimum width assert_equals: right position is 0 expected 400 but got 300
